requirements.txt
requests>=2.28.0
pycryptodome>=3.15.0
ddddocr>=1.4.0
opencv-python-headless>=4.5.0
Pillow>=9.0.0
tabulate>=0.9.0